server/requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-Cors==3.0.8
google-api-core==1.14.2
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-datastore==1.9.0
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
nltk==3.4.5
numpy==1.17.0
Pillow==6.1.0
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.3
Werkzeug==0.15.5
